requirements.txt
astroid==1.6.0
autopep8==1.3.3
Django==2.0.1
django-cors-headers==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.3.1
PyJWT==1.5.3
pylint==1.8.1
pytz==2017.3
six==1.11.0
wrapt==1.10.11
yapf==0.20.0
